feat: move from helper to util
