+++ apps/editor/src/utils/common.ts
@@ -13,11 +13,6 @@ export function includes<T>(arr: T[], targetItem: T) {
 
 const availableLinkAttributes = ['rel', 'target', 'contenteditable', 'hreflang', 'type'];
 
-/**
- * sanitize attribute for link
- * @param {?object} attribute - attribute for link
- * @returns {object} sanitized attribute
- */
 export function sanitizeLinkAttribute(attribute: Record<string, any>) {
   if (!attribute) {
     return null;
@@ -32,4 +27,36 @@ export function sanitizeLinkAttribute(attribute: Record<string, any>) {
   });
 
   return linkAttribute;
+}
+
+export function repeat(text: string, count: number) {
+  let result = '';
+
+  for (let i = 0; i < count; i += 1) {
+    result += text;
+  }
+
+  return result;
+}
+
+export function escape(text: string, startOfLine?: boolean) {
+  const result = text.replace(/[`*\\~[\]]/g, '\\$&');
+
+  if (startOfLine) {
+    return result.replace(/^[:#\-*+]/, '\\$&').replace(/^(\d+)\./, '$1\\.');
+  }
+
+  return result;
+}
+
+export function quote(text: string) {
+  let result;
+
+  if (text.indexOf('"') === -1) {
+    result = '""';
+  } else {
+    result = text.indexOf("'") === -1 ? "''" : '()';
+  }
+
+  return result[0] + text + result[1];
 }
